requirements.txt
aiofiles==0.6.0
fastapi==0.63.0
fastapi-login==1.5.3
Jinja2==2.11.3
passlib==1.7.4
pydantic==1.8.1
PyJWT==2.0.1
PyPika==0.44.1
python-multipart==0.0.5
requests==2.25.1
tortoise-orm==0.16.21
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn[standard]==0.13.4
bcrypt==3.2.0